Add support for specify test requirements of a bundle
Currently there is only limited support for "test only" dependencies of
a bundle, the usual way is to add an (optional) import/require bundle to
the test but this is not very suitable when one wants to have tests in
the bundle itself, also some requirements can't be expressed as a
package/bundle without binding to a particular implementation and even
if, these are still added as regular dependencies and not with test
scope.

This ads a way to specify a list of 'test.requirements' in the
build.properties to account for this.