[target] Avoid Vertical Scrollbars a much as possible
Due to the behavior of the used FormLayout one can easily get into a
quite annoying situation where the target editor shows two scrollbars,
one for the form page and one for the location and both compete with
each other. Even worse this often happens when resizing the view (e.g.
to see more content) and can only be reset by closing the editor
entirely.

This now uses a BorderLayout that behaves much more better in resizing
scenarios as GridLayout and avoids the two scrollbars as much as
possible.